Add headerless S3 boot script support
This type of boot script differs from the default type by omitting the
header that would contain the index and length of each entry. The length
of the entry must instead be derived from the first two bytes (opcode
and width).

A boot script of this type was found on Intel DQ67SW mainboards. With
these changes it was possible to parse the entire S3 boot script and
modify entries using `tools.uefi.s3script_modify`.